+++ apps/demo/src/demo-components/DemoProgression.tsx
@@ -18,6 +18,7 @@ import {
 } from '@digital-ai/dot-components';
 import { pbWorkItemDetailsData } from '../demo-data/pbWorkItemDetailsData';
 import { pbApplicationAPIData as apiData } from '../demo-data/pbApplicationAPIData';
+import { sampleEmptyPhases } from '../demo-data/pbEmptyPhasesData';
 import { phasesData } from './DemoData';
 import { pbAppDetailsData } from '../demo-data/pbAppDetailsData';
 
@@ -27,13 +28,16 @@ const dialogRootClassName = 'form-result-dialog';
 const StyledDemoProgression = styled.div`
   &.${rootClassName} {
     height: 100%;
+    display: flex;
+    flex-direction: column;
 
     .progression {
       display: flex;
+      height: 100%;
     }
 
-    .new-application-btn {
-      margin: 8px 0 8px 4px;
+    .progression-action {
+      margin: 8px 0 8px 0;
     }
 
     .form-child-list {
@@ -63,6 +67,9 @@ export const ProgressionDemo = () => {
     false
   );
   const [isInConfigureMode, setIsInConfigureMode] = useState<boolean>(false);
+  const [isInEmptyPhasesMode, setIsInEmptyPhasesMode] = useState<boolean>(
+    false
+  );
   const [editablePhases, setEditablePhases] = useState<
     Array<EditablePhaseType>
   >(null);
@@ -112,6 +119,10 @@ export const ProgressionDemo = () => {
     onDrawerClose();
     setIsAppDrawerOpened(true);
   };
+
+  const onViewEmptyPhasesClick = (): void => setIsInEmptyPhasesMode(true);
+
+  const onCancelEmptyPhasesMode = (): void => setIsInEmptyPhasesMode(false);
 
   const onConfigureClick = (): void => {
     setIsInConfigureMode((prevState) => !prevState);
@@ -253,17 +264,47 @@ export const ProgressionDemo = () => {
         />
       );
     }
+    if (isInEmptyPhasesMode) {
+      return (
+        <>
+          <div className="progression-action">
+            <DotButton
+              className="view-empty-phases-btn"
+              onClick={onCancelEmptyPhasesMode}
+            >
+              Go Back
+            </DotButton>
+          </div>
+          <div className="progression">
+            <DotProgressionBoard
+              className="progression-board"
+              workItemSelection={workItemSelection}
+              phases={sampleEmptyPhases}
+              baseUrl={BASE_URL}
+            />
+          </div>
+        </>
+      );
+    }
     return (
       <>
-        <DotButton
-          className="new-application-btn"
-          onClick={onNewApplicationClick}
-        >
-          New Application
-        </DotButton>
-        <DotButton className="edit-btn" onClick={onConfigureClick}>
-          Configure
-        </DotButton>
+        <div className="progression-action">
+          <DotButton
+            className="new-application-btn"
+            onClick={onNewApplicationClick}
+          >
+            New Application
+          </DotButton>
+          <DotButton
+            className="view-empty-phases-btn"
+            onClick={onViewEmptyPhasesClick}
+          >
+            View Empty Phases
+          </DotButton>
+          <DotButton className="edit-btn" onClick={onConfigureClick}>
+            Configure
+          </DotButton>
+        </div>
         <div className="progression">
           <DotProgressionBoard
             className="progression-board"