+++ libs/dot-components/src/lib/ctm/progression-board/ProgressionBoard.styles.ts
@@ -16,6 +16,8 @@ export const StyledTooltipContent = styled(DotTypography)`
 export const StyledProgressionBoard = styled.div<ProgressionBoardProps>`
   ${({ theme, offsetLeft }) => css`
     &.${rootClassName} {
+      height: 100%;
+
       &.columns-wrapper {
         display: table;
         flex: 1;