+++ libs/dot-components/src/lib/ctm/progression-board/ProgressionBoard.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useRef } from 'react';
+import React, { useEffect, useState, useRef, ReactNode } from 'react';
 import { CommonProps } from '../../components/CommonProps';
 import { useStylesWithRootClass } from '../../components/useStylesWithRootClass';
 import { BoardHeaders } from './BoardHeaders';
@@ -8,12 +8,14 @@ import {
 } from './ProgressionBoard.styles';
 import { calculateProgressionBoardOffset } from './progressionBoardHelper';
 import {
+  PackageType,
   PhaseType,
   SelectedWorkItem,
   SwimLanepkg,
   WorkItemSelection,
 } from './ProgressionBoardInterfaces';
 import { SwimLane } from './SwimLane';
+import { EmptyPhases } from './phase/EmptyPhases';
 
 export interface ProgressionBoardProps extends CommonProps {
   /* Base URL on which user will be redirected when item selection is made */
@@ -93,8 +95,8 @@ export const DotProgressionBoard = ({
     // in which those workitems appear.
     const workitemMap: WorkitemMapType = {};
     phases.forEach((phase) => {
-      phase.packageVersions.forEach((packageVersion) => {
-        packageVersion.workitems.forEach((workitem) => {
+      phase.packageVersions.forEach((packageVersion: PackageType) => {
+        packageVersion.workitems?.forEach((workitem) => {
           const currentCount = workitemMap[workitem._id] || 0;
           workitemMap[workitem._id] = currentCount + 1;
         });
@@ -106,12 +108,12 @@ export const DotProgressionBoard = ({
         // while here, use the map we built above to determine
         // if workitems are split.
         .map((phase) => {
-          phase.packageVersions.forEach((packageVersion) => {
-            packageVersion.workitems.forEach((workitem) => {
+          phase.packageVersions.forEach((packageVersion: PackageType) => {
+            packageVersion.workitems?.forEach((workitem) => {
               workitem.isSplit = workitemMap[workitem._id] > 1;
             });
           });
-          return phase.packageVersions.map((version) => ({
+          return phase.packageVersions.map((version: PackageType) => ({
             package_id: version.package_id,
             package_name: version.package_name,
           }));
@@ -141,7 +143,7 @@ export const DotProgressionBoard = ({
           const packagePhases = phases.map((phase) => ({
             ...phase,
             packageVersions: phase.packageVersions.filter(
-              (version) => version.package_id === pkg.package_id
+              (version: PackageType) => version.package_id === pkg.package_id
             ),
           }));
 
@@ -152,6 +154,8 @@ export const DotProgressionBoard = ({
         })
     );
   };
+
+  const allPackages = getPackages();
 
   const pbRef = useRef(null);
 
@@ -168,6 +172,23 @@ export const DotProgressionBoard = ({
     ));
   };
 
+  const renderProgressionBoard = (): ReactNode => {
+    if (allPackages && allPackages.length > 0) {
+      return (
+        <>
+          <BoardHeaders headers={phaseNames} isOffsetLeft={offsetLeft > 0} />
+          {renderSwimLanesFromPackages(allPackages)}
+        </>
+      );
+    }
+    return (
+      <EmptyPhases
+        data-testid={`${dataTestId}-empty-phases`}
+        phaseNames={phaseNames}
+      />
+    );
+  };
+
   return (
     <StyledProgressionBoard
       className={rootClasses}
@@ -176,8 +197,7 @@ export const DotProgressionBoard = ({
       offsetLeft={offsetLeft}
       ref={pbRef}
     >
-      <BoardHeaders headers={phaseNames} isOffsetLeft={offsetLeft > 0} />
-      {renderSwimLanesFromPackages(getPackages())}
+      {renderProgressionBoard()}
     </StyledProgressionBoard>
   );
 };